[update]メニューの設定実装

+++ Assets/Scripts/DataDriven/ApplicationLayer/System/PlaySceneSystem.cs
@@ -1,5 +1,6 @@
 using System;
 using UnityEngine;
+
 namespace DataDriven
 {
     /// <summary>プレイ中の処理を司るクラス</summary>
@@ -18,7 +19,7 @@ namespace DataDriven
         /// <param name="index">選択するスロット</param>
         public void HotbarSelectForKetboard(int index)
         {
-            if (_repository.TryGetData<HotbarRuntimeData>((int)DataID.Hotbar, out var hotbar))
+            if (_repository.TryGetData<HotbarRuntimeData>(DataID.Hotbar, out var hotbar))
             {
                 hotbar.SelectItemForKeyboard(index);
             }
@@ -30,7 +31,7 @@ namespace DataDriven
         /// <param name="dir">選択するスロットをずらす方向</param>
         public void HotbarSelectForGamePad(IndexMove dir)
         {
-            if (_repository.TryGetData<HotbarRuntimeData>((int)DataID.Hotbar, out var hotbar))
+            if (_repository.TryGetData<HotbarRuntimeData>(DataID.Hotbar, out var hotbar))
             {
                 hotbar.SelectItemForGamePad(dir);
             }
@@ -41,7 +42,7 @@ namespace DataDriven
         /// </summary>
         public void UseItem()
         {
-            if (_repository.TryGetData<HotbarRuntimeData>((int)DataID.Hotbar, out var hotbar))
+            if (_repository.TryGetData<HotbarRuntimeData>(DataID.Hotbar, out var hotbar))
             {
                 var item = hotbar.UseItem();
                 //空のスロットを選択した時
